requirements.txt
Flask
Flask-CORS
supabase
lunar-python
python-dotenv
bcrypt
PyJWT